fix: correct import paths in _components folders after directory rename
- Fixed all component imports to use './' for same-directory imports
- Retained '../' for actual parent directory resources (data/constants, utils)
- Removed unused test imports (vi, beforeEach, afterEach)
- Fixed TypeScript type assertion in use-project-tabs.ts
- All lint and type-check passing
