+++ src/hooks/__tests__/use-local-storage.test.ts
@@ -1,7 +1,7 @@
 /**
  * Unit tests for useLocalStorage hook
  */
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
+import { describe, it, expect, vi, afterEach } from 'vitest'
 import { renderHook, act } from '@testing-library/react'
 import { useLocalStorage } from '../use-local-storage'
